+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Correctly encode Infinity, -Infinity and NaN in json
 * Set stepIgnore on log hook functions
 * Support new metamethod `__debugvisualize(self)` for formatting values for [hediet.debug-visualizer](https://marketplace.visualstudio.com/items?itemName=hediet.debug-visualizer)
+* Debug line and visualizer for noise expressions
 
 ## 0.18.22